added exception and swagger

+++ src/main/java/com/example/myappDB/controller/CustomerController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.myappDB.entities.Customer;
+import com.example.myappDB.exception.ApiRequestException;
 import com.example.myappDB.service.CustomerRepository;
 
 @RestController
@@ -28,39 +29,45 @@ public class CustomerController {
 	
 	@GetMapping("/customers/{id}")
 	public Customer getCustomerById(@PathVariable Integer id) {
-//		Optional<Customer> val = customerService.findById(id);
-//
-//        if (val.isPresent()) {
-//            System.out.println(val.get());
-//        } else {
-//            System.out.printf("No customer found with id %d%n", id);
-//        }
-		return customerService.findById(id).get();
+		Optional<Customer> customer = customerService.findById(id);
+		if (customer.isPresent()) {
+            return customer.get();
+        } else {
+        	throw new ApiRequestException("No customer found with id " + id);
+        }
 	}
 	
 	@PostMapping("/customers")
 	public String addNewCustomer(@RequestBody Customer customer) {
 		customerService.save(customer);
-		return "Customer " + customer.getId() + " added";
+		return "New Customer with id " + customer.getId() + " is added";
 	} 
 	
 	@PutMapping("/customers/{id}")
 	public String updateCustomer(@PathVariable Integer id, @RequestBody Customer customerDetails) {
+				
+		Optional<Customer> customerToUpdate = customerService.findById(id);
+		if (!customerToUpdate.isPresent()) {
+			customerService.save(customerDetails);
+			return "New Customer with id " + customerDetails.getId() + " is added";
+//			throw new ApiRequestException("No customer found with id " + id);
+        } 
 		
-		Customer customerToUpdate = customerService.findById(id).get();
+		customerToUpdate.get().setFirstName(customerDetails.getFirstName());
+		customerToUpdate.get().setLastName(customerDetails.getLastName());
 		
-		customerToUpdate.setFirstName(customerDetails.getFirstName());
-		customerToUpdate.setLastName(customerDetails.getLastName());
-		
-		customerService.save(customerToUpdate);
-		
-		return "Customer " + id + " updated";
+		customerService.save(customerToUpdate.get());
+		return "Customer with id " + id + " is updated";
 	}
 	
 	@DeleteMapping("/customers/{id}")
 	public String deleteCustomerById(@PathVariable Integer id) {
+		Optional<Customer> customerToDelete = customerService.findById(id);
+		if (!customerToDelete.isPresent()) {
+			throw new ApiRequestException("No customer found with id " + id);
+        } 
 		customerService.deleteById(id);
-		return "Customer " + id + " deleted";
+		return "Customer with id " + id + " is deleted";
 	}
